Add client functionality to bitcoin rpc clients (#2632)
* refactor: clean up some of the bitcoin rpc code

* feature: standardize on get_tx_block_height

* refactor: split out clients into features
