+++ modules/fedimint-wallet-client/Cargo.toml
@@ -18,6 +18,7 @@ async-trait = "0.1"
 bitcoin = { version = "0.29.2", features = [ "rand", "serde"] }
 erased-serde = "0.3"
 esplora-client = { version = "0.5.0", default-features = false, features = ["async", "async-https-rustls"] }
+fedimint-bitcoind = { path = "../../fedimint-bitcoind", default-features = false, features = ["esplora-client", "bitcoincore-rpc"] }
 fedimint-client = { path = "../../fedimint-client" }
 fedimint-core = { path = "../../fedimint-core" }
 fedimint-wallet-common ={ path = "../fedimint-wallet-common" }
